Move notification updates on session into a job
With jobs, I seem to have more control over when they should run compared to having checks in the session flow. This way I don't have to check if I'm bound or a config change happened, I just need to make sure the job is cancelled at the right time to stop collecting.